Move processed CSS out. Add brand colors.
