+++ package.json
@@ -34,8 +34,8 @@
     "react-test-renderer": "^15.5.4"
   },
   "scripts": {
-    "build-css": "node-sass src/ -o src/",
-    "watch-css": "npm run build-css && node-sass src/ -o src/ --watch --recursive",
+    "build-css": "node-sass src/ -o public/",
+    "watch-css": "npm run build-css && node-sass src/ -o public/ --watch --recursive",
     "start-js": "react-scripts-ts start",
     "start": "npm-run-all -p watch-css start-js",
     "build": "react-scripts-ts build && npm run build-css",
