+++ public/stylesheets/App.css
@@ -1,0 +1,26 @@
+.App {
+  text-align: center;
+  font-family: "Palatino Linotype", "Book Antiqua", Palatino, serif;
+  color: #4C4845; }
+
+.app-logo {
+  height: 50px;
+  padding: 20px;
+  float: left; }
+
+.app-header {
+  background-color: #222;
+  height: 90px;
+  color: white; }
+  .app-header div {
+    display: block;
+    float: left; }
+
+.app-title {
+  text-align: left; }
+  .app-title h2 {
+    padding-top: 2px;
+    margin-bottom: 4px; }
+  .app-title span {
+    font-size: 10pt;
+    vertical-align: top; }
