--- src/stylesheets/App.css
@@ -1,19 +1,0 @@
-.App {
-  text-align: center;
-}
-
-.App-logo {
-  height: 80px;
-  float: left;
-}
-
-.App-header {
-  background-color: #222;
-  height: 100px;
-  padding: 20px;
-  color: white;
-}
-
-.App-intro {
-  font-size: large;
-}
