+++ src/stylesheets/App.scss
@@ -1,0 +1,35 @@
+.App {
+  text-align: center;
+  font-family: "Palatino Linotype", "Book Antiqua", Palatino, serif;
+  color: #4C4845; 
+}
+
+.app-logo {
+  height: 50px;
+  padding: 20px;
+  float: left; }
+
+.app-header {
+  background-color: #222;
+  height: 90px;
+  color: white;
+
+  div {
+    display: block;
+    float: left;
+  }
+}
+
+.app-title {
+  text-align: left;
+
+  h2 {
+    padding-top: 2px;
+    margin-bottom: 4px;
+  }
+
+  span {
+    font-size: 10pt;
+    vertical-align: top;
+  }
+}
