+++ src/stylesheets/Slideshow.scss
@@ -1,3 +1,8 @@
+@import "./colors";
+
+$button-background: #A5488B;
+$button-background-highlighted: #7F386B;
+
 .slideshow-container {
   width: 50%;
   margin-left: 25%;
@@ -21,37 +26,37 @@
   text-decoration: none;
   outline: none;
   color: #fff;
-  background-color: #4CAF50;
+  background-color: $button-background; 
   border: none;
   border-radius: 12px;
   box-shadow: 0 8px #999;
   margin-top: 8px;
 
   &:active {
-    background-color: #3e8e41;
+    background-color: $button-background-highlighted;
     box-shadow: 0 5px #666;
     transform: translateY(4px);
   }
 
   &:hover {
-    background-color: #3e8e41;
+    background-color: $button-background-highlighted;
   }
 }
 
 .next-button {
   @extend .directional-button;
-  background-color: #337AB7;
+  background-color: $teal;
   float: right;
 
   &:hover {
-    background-color: #286090; 
+    background-color: $teal-highlighted; 
   }
 }
 
 .previous-button {
   @extend .directional-button;
   float: left;
-  background-color: #FFF;
+  background-color: white;
   border: 1px solid #CCC;
   color: #333;
