+++ src/stylesheets/colors.scss
@@ -1,0 +1,4 @@
+$black: #4C4845;
+$red: #FD4F57;
+$teal: #16CBC4;
+$teal-highlighted: #0E7F7B;
